# contracts/openapi/public-service.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InputSourceCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "poll_interval_seconds": {
            "default": 900,
            "maximum": 86400.0,
            "minimum": 30.0,
            "title": "Poll Interval Seconds",
            "type": "integer"
          },
          "provider": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Provider",
            "type": "string"
          },
          "secrets": {
            "additionalProperties": true,
            "title": "Secrets",
            "type": "object"
          },
          "source_key": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Key"
          },
          "source_kind": {
            "enum": [
              "calendar",
              "email",
              "task",
              "exam",
              "announcement"
            ],
            "title": "Source Kind",
            "type": "string"
          }
        },
        "required": [
          "source_kind",
          "provider"
        ],
        "title": "InputSourceCreateRequest",
        "type": "object"
      },
      "InputSourcePatchRequest": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "poll_interval_seconds": {
            "anyOf": [
              {
                "maximum": 86400.0,
                "minimum": 30.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poll Interval Seconds"
          },
          "secrets": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secrets"
          }
        },
        "title": "InputSourcePatchRequest",
        "type": "object"
      },
      "InputSourceResponse": {
        "properties": {
          "active_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Request Id"
          },
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "config_state": {
            "enum": [
              "stable",
              "rebind_pending"
            ],
            "title": "Config State",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "last_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Code"
          },
          "last_error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error Message"
          },
          "last_polled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Polled At"
          },
          "lifecycle_state": {
            "enum": [
              "active",
              "inactive",
              "archived"
            ],
            "title": "Lifecycle State",
            "type": "string"
          },
          "next_poll_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Poll At"
          },
          "oauth_account_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oauth Account Email"
          },
          "oauth_connection_status": {
            "anyOf": [
              {
                "enum": [
                  "connected",
                  "not_connected"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oauth Connection Status"
          },
          "operator_guidance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceOperatorGuidanceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "poll_interval_seconds": {
            "title": "Poll Interval Seconds",
            "type": "integer"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "runtime_state": {
            "enum": [
              "active",
              "inactive",
              "archived",
              "queued",
              "running",
              "rebind_pending"
            ],
            "title": "Runtime State",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "integer"
          },
          "source_key": {
            "title": "Source Key",
            "type": "string"
          },
          "source_kind": {
            "enum": [
              "calendar",
              "email",
              "task",
              "exam",
              "announcement"
            ],
            "title": "Source Kind",
            "type": "string"
          },
          "sync_progress": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sync Progress"
          },
          "sync_state": {
            "enum": [
              "idle",
              "queued",
              "running"
            ],
            "title": "Sync State",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "source_id",
          "user_id",
          "source_kind",
          "provider",
          "source_key",
          "display_name",
          "is_active",
          "poll_interval_seconds",
          "last_polled_at",
          "next_poll_at",
          "last_error_code",
          "last_error_message",
          "created_at",
          "updated_at",
          "config",
          "lifecycle_state",
          "sync_state",
          "config_state",
          "runtime_state"
        ],
        "title": "InputSourceResponse",
        "type": "object"
      },
...
      "SourceObservabilityResponse": {
        "properties": {
          "active": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceObservabilitySyncResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Request Id"
          },
          "bootstrap": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceObservabilitySyncResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "latest_replay": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceObservabilitySyncResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "operator_guidance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceOperatorGuidanceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "source_id": {
            "title": "Source Id",
            "type": "integer"
          }
        },
        "required": [
          "source_id"
        ],
        "title": "SourceObservabilityResponse",
        "type": "object"
      },
      "SourceObservabilitySyncResponse": {
        "properties": {
          "applied": {
            "default": false,
            "title": "Applied",
            "type": "boolean"
          },
          "applied_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Applied At"
          },
          "connector_result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connector Result"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Ms"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "llm_usage": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Usage"
          },
          "phase": {
            "enum": [
              "bootstrap",
              "replay"
            ],
            "title": "Phase",
            "type": "string"
          },
          "progress": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "stage": {
            "anyOf": [
              {
                "enum": [
                  "connector_fetch",
                  "llm_queue",
                  "llm_parse",
                  "provider_reduce",
                  "result_ready",
                  "applying",
                  "completed",
                  "failed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "stage_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage Updated At"
          },
          "status": {
            "enum": [
              "PENDING",
              "QUEUED",
              "RUNNING",
              "SUCCEEDED",
              "FAILED"
            ],
            "title": "Status",
            "type": "string"
          },
          "substage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Substage"
          },
          "trigger_type": {
            "enum": [
              "manual",
              "scheduler",
              "webhook"
            ],
            "title": "Trigger Type",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "phase",
          "trigger_type",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "SourceObservabilitySyncResponse",
        "type": "object"
      },
      "SourceOperatorGuidanceResponse": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "progress_age_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress Age Seconds"
          },
          "reason_code": {
            "title": "Reason Code",
            "type": "string"
          },
          "recommended_action": {
            "enum": [
              "continue_review",
              "continue_review_with_caution",
              "wait_for_runtime",
              "investigate_runtime"
            ],
            "title": "Recommended Action",
            "type": "string"
          },
          "related_request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Related Request Id"
          },
          "severity": {
            "enum": [
              "info",
              "warning",
              "blocking"
            ],
            "title": "Severity",
            "type": "string"
          }
        },
        "required": [
          "recommended_action",
          "severity",
          "reason_code",
          "message"
        ],
        "title": "SourceOperatorGuidanceResponse",
        "type": "object"
      },
      "SourceSyncHistoryResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SourceObservabilitySyncResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "source_id": {
            "title": "Source Id",
            "type": "integer"
          }
        },
        "required": [
          "source_id",
          "items"
        ],
        "title": "SourceSyncHistoryResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "CalendarDIFF API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/sources": {
      "get": {
        "operationId": "list_sources_sources_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "active",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InputSourceResponse"
                  },
                  "title": "Response List Sources Sources Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sources",
        "tags": [
          "input-control-plane"
        ]
      },
      "post": {
        "operationId": "create_source_sources_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InputSourceCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputSourceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Source",
        "tags": [
          "input-control-plane"
        ]
      }
    },
    "/sources/{source_id}": {
      "delete": {
        "operationId": "delete_source_sources__source_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Delete Source Sources  Source Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Source",
        "tags": [
          "input-control-plane"
        ]
      },
      "patch": {
        "operationId": "patch_source_sources__source_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InputSourcePatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputSourceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Source",
        "tags": [
          "input-control-plane"
        ]
      }
    },
...
    "/sources/{source_id}/observability": {
      "get": {
        "operationId": "get_source_observability_sources__source_id__observability_get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceObservabilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Source Observability",
        "tags": [
          "input-control-plane"
        ]
      }
    },
    "/sources/{source_id}/sync-history": {
      "get": {
        "operationId": "get_source_sync_history_sources__source_id__sync_history_get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceSyncHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Source Sync History",
        "tags": [
          "input-control-plane"
        ]
      }
    },
...
  }
}